+++ src/ui/server/controllers/insurance/export-contract/how-was-the-contract-awarded/index.test.ts
@@ -10,7 +10,8 @@ import getUserNameFromSession from '../../../../helpers/get-user-name-from-sessi
 import constructPayload from '../../../../helpers/construct-payload';
 import generateValidationErrors from './validation';
 import { sanitiseData } from '../../../../helpers/sanitise-data';
-import { Request, Response } from '../../../../../types';
+import mapAndSave from '../map-and-save/export-contract';
+import { ObjectType, Request, Response } from '../../../../../types';
 import { mockReq, mockRes, mockExportContract, referenceNumber } from '../../../../test-mocks';
 
 const {
@@ -34,6 +35,10 @@ const {
 describe('controllers/insurance/export-contract/how-was-the-contract-awarded', () => {
   let req: Request;
   let res: Response;
+
+  jest.mock('../map-and-save/export-contract');
+
+  mapAndSave.exportContract = jest.fn(() => Promise.resolve(true));
 
   beforeEach(() => {
     req = mockReq();
@@ -98,6 +103,7 @@ describe('controllers/insurance/export-contract/how-was-the-contract-awarded', (
         ...pageVariables(),
         CONDITIONAL_OTHER_METHOD_HTML,
         userName: getUserNameFromSession(req.session.user),
+        submittedValues: mockExportContract,
       };
 
       expect(res.render).toHaveBeenCalledWith(TEMPLATE, expectedVariables);
@@ -127,6 +133,8 @@ describe('controllers/insurance/export-contract/how-was-the-contract-awarded', (
 
         const payload = constructPayload(req.body, FIELD_IDS);
 
+        const sanitised = sanitiseData(payload) as ObjectType;
+
         const expectedVariables = {
           ...insuranceCorePageVariables({
             PAGE_CONTENT_STRINGS,
@@ -135,7 +143,12 @@ describe('controllers/insurance/export-contract/how-was-the-contract-awarded', (
           ...pageVariables(),
           CONDITIONAL_OTHER_METHOD_HTML,
           userName: getUserNameFromSession(req.session.user),
-          submittedValues: sanitiseData(payload),
+          submittedValues: {
+            ...sanitised,
+            [AWARD_METHOD]: {
+              id: sanitised[AWARD_METHOD],
+            },
+          },
           validationErrors: generateValidationErrors(payload),
         };
 
@@ -146,6 +159,18 @@ describe('controllers/insurance/export-contract/how-was-the-contract-awarded', (
     describe('when there are no validation errors', () => {
       beforeEach(() => {
         req.body = validBody;
+      });
+
+      it('should call mapAndSave.exportContract with data from constructPayload function and application', async () => {
+        await post(req, res);
+
+        const payload = constructPayload(req.body, FIELD_IDS);
+
+        expect(mapAndSave.exportContract).toHaveBeenCalledTimes(1);
+
+        const expectedValidationErrors = false;
+
+        expect(mapAndSave.exportContract).toHaveBeenCalledWith(payload, res.locals.application, expectedValidationErrors);
       });
 
       it(`should redirect to ${ABOUT_GOODS_OR_SERVICES}`, async () => {
@@ -160,6 +185,36 @@ describe('controllers/insurance/export-contract/how-was-the-contract-awarded', (
     describe('when there is no application', () => {
       beforeEach(() => {
         delete res.locals.application;
+      });
+
+      it(`should redirect to ${PROBLEM_WITH_SERVICE}`, async () => {
+        await post(req, res);
+
+        expect(res.redirect).toHaveBeenCalledWith(PROBLEM_WITH_SERVICE);
+      });
+    });
+
+    describe('when mapAndSave.exportContract does not return a true boolean', () => {
+      beforeEach(() => {
+        req.body = validBody;
+        const mapAndSaveSpy = jest.fn(() => Promise.resolve(false));
+
+        mapAndSave.exportContract = mapAndSaveSpy;
+      });
+
+      it(`should redirect to ${PROBLEM_WITH_SERVICE}`, async () => {
+        await post(req, res);
+
+        expect(res.redirect).toHaveBeenCalledWith(PROBLEM_WITH_SERVICE);
+      });
+    });
+
+    describe('when mapAndSave.exportContract returns an error', () => {
+      beforeEach(() => {
+        req.body = validBody;
+        const mapAndSaveSpy = jest.fn(() => Promise.reject(new Error('mock')));
+
+        mapAndSave.exportContract = mapAndSaveSpy;
       });
 
       it(`should redirect to ${PROBLEM_WITH_SERVICE}`, async () => {