+++ src/ui/server/controllers/insurance/export-contract/map-submitted-data/export-contract/index.test.ts
@@ -1,14 +1,16 @@
 import mapSubmittedData from '.';
 import FIELD_IDS from '../../../../../constants/field-ids/insurance/export-contract';
 import getCountryByIsoCode from '../../../../../helpers/get-country-by-iso-code';
-import { mockCountries } from '../../../../../test-mocks';
+import { mockCountries, mockExportContract } from '../../../../../test-mocks';
 import { RequestBody } from '../../../../../../types';
 
 const {
+  HOW_WAS_THE_CONTRACT_AWARDED: { AWARD_METHOD },
   ABOUT_GOODS_OR_SERVICES: { DESCRIPTION, FINAL_DESTINATION_KNOWN, FINAL_DESTINATION },
 } = FIELD_IDS;
 
 const mockCountryIsoCode = mockCountries[0].isoCode;
+const mockAwardMethodId = mockExportContract.awardMethod.id;
 
 describe('controllers/insurance/export-contract/map-submitted-data/export-contract', () => {
   let mockFormBody: RequestBody;
@@ -18,6 +20,28 @@ describe('controllers/insurance/export-contract/map-submitted-data/export-contra
       [DESCRIPTION]: 'Mock description about my services',
       [FINAL_DESTINATION_KNOWN]: 'true',
     };
+  });
+
+  describe(`when ${AWARD_METHOD} field is provided`, () => {
+    it(`should return the form body with ${FINAL_DESTINATION} as a country ISO code`, () => {
+      const mockBodyWithAwardMethod = {
+        ...mockFormBody,
+        [AWARD_METHOD]: mockAwardMethodId,
+      };
+
+      const result = mapSubmittedData(mockBodyWithAwardMethod);
+
+      const expected = {
+        ...mockFormBody,
+        [AWARD_METHOD]: {
+          connect: {
+            id: mockAwardMethodId,
+          },
+        },
+      };
+
+      expect(result).toEqual(expected);
+    });
   });
 
   describe(`when ${FINAL_DESTINATION} field is provided`, () => {